+++ SMLHelper/Crafting/Node.cs
@@ -1,0 +1,14 @@
+﻿namespace SMLHelper.V2.Crafting
+{
+    internal class Node
+    {
+        internal string[] Path { get; set; }
+        internal CraftTree.Type Scheme { get; set; }
+
+        internal Node(string[] path, CraftTree.Type scheme)
+        {
+            Path = path;
+            Scheme = scheme;
+        }
+    }
+}
