+++ SMLHelper/SMLHelper.csproj
@@ -69,9 +69,16 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Crafting\NodeFamily.cs" />
+    <Compile Include="Crafting\CraftingNode.cs" />
+    <Compile Include="Crafting\ModCraftTreeCraft.cs" />
+    <Compile Include="Crafting\ModCraftTreeLinkingNode.cs" />
+    <Compile Include="Crafting\ModCraftTreeNode.cs" />
+    <Compile Include="Crafting\ModCraftTreeRoot.cs" />
+    <Compile Include="Crafting\ModCraftTreeTab.cs" />
     <Compile Include="Assets\CustomFish.cs" />
     <Compile Include="Assets\CustomFishPrefab.cs" />
+    <Compile Include="Crafting\Node.cs" />
+    <Compile Include="Crafting\TabNode.cs" />
     <Compile Include="Patchers\CustomFishPatcher.cs" />
     <Compile Include="Handlers\CustomFishHandler.cs" />
     <Compile Include="Handlers\ItemActionHandler.cs" />
@@ -97,7 +104,6 @@
     <Compile Include="Legacy\CustomCraftTreeFamily.cs" />
     <Compile Include="Assets\ModPrefab.cs" />
     <Compile Include="Assets\ModSprite.cs" />
-    <Compile Include="Crafting\ModCraftTreeFamily.cs" />
     <Compile Include="Crafting\TechData.cs" />
     <Compile Include="Handlers\BioReactorHandler.cs" />
     <Compile Include="Handlers\CraftDataHandler.cs" />
